feat: Add token exchange authentication provider
This commit introduces a new authentication mechanism that allows the OCI TypeScript SDK to authenticate with OCI IAM Domains by exchanging a third-party token for a UPST token.

The new  enables this functionality. It can be configured using environment variables or programmatically with a builder pattern.

A new example, , has been added to demonstrate how to use the new provider.

This change enables customers to leverage their existing identity providers to authenticate with OCI services.

Signed-off-by: Gordon Trevorrow <gordon.trevorrow@oracle.com>